refactor(view): 폴더 구조 수정 - InputView, OutputView `console`폴더로 분류
